+++ R/dashboardPage.R
@@ -51,37 +51,23 @@ bs4DashPage <- function(navbar = NULL, sidebar = NULL, body = NULL,
       ),
       shiny::tags$meta(`http-equiv` = "x-ua-compatible", content = "ie=edge"),
 
-      shiny::tags$title(title),
-
-      # icons
-      shiny::includeCSS(system.file("css/font-awesome.min.css", package = "bs4Dash")),
-      shiny::tags$link(href = "https://use.fontawesome.com/releases/v5.0.13/css/all.css", rel = "stylesheet"),
-      shiny::tags$link(href = "https://unpkg.com/ionicons@4.4.2/dist/css/ionicons.min.css", rel = "stylesheet"),
-      # AdminLTE3
-      shiny::includeCSS(system.file("css/adminlte.min.css", package = "bs4Dash")),
-      
-      # Old school design (bootswatch sketchy design)
-      if (old_school) {
-        shiny::tags$link(href = "https://bootswatch.com/4/sketchy/bootstrap.min.css", rel = "stylesheet")
-      },
-      shiny::tags$link(href = "https://fonts.googleapis.com/css?family=Source+Sans+Pro:300,400,400i,700", rel = "stylesheet")
+      shiny::tags$title(title)
     ),
     # Body
-    shiny::tags$body(
-      class = paste0("hold-transition sidebar-mini ", sidebar_cl),
-      shiny::tags$div(
-        class = "wrapper",
-        navbar,
-        sidebar,
-        # page content
-        body,
-        controlbar,
-        footer
-      ),
-      # javascript
-      shiny::includeScript(system.file("js/bootstrap.bundle.min.js", package = "bs4Dash")),
-      shiny::includeScript(system.file("js/adminlte.min.js", package = "bs4Dash")),
-      shiny::includeScript(system.file("js/bs4Dash.js", package = "bs4Dash"))
+    addDeps(
+      theme = old_school,
+      shiny::tags$body(
+        class = paste0("hold-transition sidebar-mini ", sidebar_cl),
+        shiny::tags$div(
+          class = "wrapper",
+          navbar,
+          sidebar,
+          # page content
+          body,
+          controlbar,
+          footer
+        )
+      )
     )
   )
 }